% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbars.R, R/components.R
\docType{methods}
\name{colorbars,Iheatmap-method}
\alias{colorbars,Iheatmap-method}
\alias{access_component}
\alias{yaxes,Iheatmap-method}
\alias{xaxes,Iheatmap-method}
\alias{plots,Iheatmap-method}
\alias{shapes,Iheatmap-method,}
\alias{annotations,Iheatmap-method}
\alias{yaxes}
\alias{xaxes}
\alias{plots}
\alias{shapes}
\alias{annotations}
\alias{colorbars}
\alias{colorbars,IheatmapColorbars-method}
\alias{colorbars,IheatmapPlots-method}
\alias{shapes,Iheatmap-method}
\title{Access subcomponents of Iheatmap object}
\usage{
\S4method{colorbars}{Iheatmap}(x, what = c("all", "continuous", "discrete"))

\S4method{yaxes}{Iheatmap}(p, xaxis = NULL)

\S4method{xaxes}{Iheatmap}(p, yaxis = NULL)

\S4method{plots}{Iheatmap}(x)

\S4method{shapes}{Iheatmap}(x)

\S4method{annotations}{Iheatmap}(x)
}
\arguments{
\item{x}{\code{\link{Iheatmap-class}} object}

\item{xaxis}{name of xaxis}

\item{yaxis}{name of yaxis}
}
\description{
These are methods for accessing subcomponents of the Iheatmap object
}
\keyword{internal}
