% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\docType{methods}
\name{add_col_signal}
\alias{add_col_signal}
\alias{add_col_signal,Iheatmap-method}
\title{add_col_signal}
\usage{
\S4method{add_col_signal}{Iheatmap}(p, signal, name, title = name,
  yname = NULL, xname = current_xaxis(p), pname = name,
  colorbar_position = get_colorbar_position(p),
  colors = pick_continuous_colors(zmid, zmin, zmax, p = p),
  zmin = min(signal, na.rm = TRUE), zmax = max(signal, na.rm = TRUE),
  zmid = 0, side = c("top", "bottom"), size = 0.05, buffer = 0.015,
  show_colorbar = TRUE, show_title = TRUE, layout = list())
}
\arguments{
\item{p}{iheatmap object}

\item{signal}{vector of signal}

\item{name}{name of colorbar}

\item{title}{label for y axis}

\item{yname}{internal name of yaxis}

\item{xname}{internal name of xaxis}

\item{pname}{internal name of plot}

\item{colorbar_position}{colorbar placement}

\item{colors}{palette or vector of colors to use}

\item{zmin}{minimum for colorscale}

\item{zmax}{maximum for colorscale}

\item{zmid}{midpoint for colorscale}

\item{side}{side of plot on which to add groups}

\item{size}{relative size of dendrogram (relative to the main heatmap)}

\item{buffer}{amount of space to leave empty before this plot, relative to size 
of first heatmap}

\item{show_colorbar}{show the colorbar?}

\item{show_title}{show title as axis label}

\item{layout}{y axis layout parameters to use}
}
\value{
\code{\link{Iheatmap-class}} object, which can be printed to generate 
an interactive graphic
}
\description{
Adds column signal to iheatmap object
}
\examples{

mat <- matrix(rnorm(20), ncol = 5, nrow = 4)  
iheatmap(mat) \%>\% add_col_signal(signal = 1:5, name = "Strength")
}
\seealso{
\code{\link{iheatmap}}, \code{\link{as_plotly}}, 
\code{\link{add_row_groups}}

\code{\link{add_row_signal}}, \code{\link{iheatmap}}, 
\code{\link{add_col_annotation}}
}
\author{
Alicia Schep
}
