% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_plotly.R
\docType{methods}
\name{as_plotly}
\alias{as_plotly}
\alias{as_plotly,Iheatmap-method}
\title{as_plotly}
\usage{
\S4method{as_plotly}{Iheatmap}(p, source = "Iheatmap")
}
\arguments{
\item{p}{\code{\link{Iheatmap-class}} object to convert}

\item{source}{source name for plot, for use in shiny}
}
\value{
plotly htmlwidgets object
}
\description{
Function to convert \code{link{Iheatmap-class}} object to plotly widget object
}
\examples{

mat <- matrix(rnorm(24), nrow = 6)
hm <- iheatmap(mat) \%>\% as_plotly()
class(hm)

# Print heatmap if interactive session 
if (interactive()) hm 
}
\seealso{
\code{\link{iheatmap}}, \code{\link{main_heatmap}}
}
\author{
Alicia Schep
}
