% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ColumnAnnotation-class}
\alias{ColumnAnnotation-class}
\alias{ColumnAnnotation}
\title{ColumnAnnotation}
\description{
Class for storing row annotation
}
\section{Slots}{

\describe{
\item{\code{xaxis}}{name of xaxis}

\item{\code{yaxis}}{name of yaxis}

\item{\code{data}}{vector of annotation values}

\item{\code{colorbar}}{name of colorbar}

\item{\code{show_colorbar}}{show the colorbar?}
}}

\seealso{
\code{\link{add_col_annotation}}, \code{\link{add_col_signal}},
\code{\link{add_col_groups}}
}
\author{
Alicia Schep
}
\keyword{internal}
