% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ColumnLabels-class}
\alias{ColumnLabels-class}
\alias{ColumnLabels}
\title{ColumnLabels}
\description{
Class for storing column labels
}
\section{Slots}{

\describe{
\item{\code{xaxis}}{name of xaxis}

\item{\code{yaxis}}{name of yaxis}

\item{\code{data}}{the names of the tick labels}

\item{\code{positions}}{the positions of the tick labels}

\item{\code{side}}{side of plot on which dendrogram is positioned, controls 
orientation}

\item{\code{textangle}}{angle for text}

\item{\code{font}}{list of font attributes}
}}

\seealso{
\code{\link{add_col_labels}}
}
\author{
Alicia Schep
}
\keyword{internal}
