% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ColumnTitle-class}
\alias{ColumnTitle-class}
\alias{ColumnTitle}
\title{ColumnTitle}
\description{
Class for storing column title
}
\section{Slots}{

\describe{
\item{\code{xaxis}}{name of xaxis}

\item{\code{yaxis}}{name of yaxis}

\item{\code{data}}{the title (character)}

\item{\code{side}}{side of plot on which dendrogram is positioned, controls 
orientation}

\item{\code{textangle}}{angle for text}

\item{\code{font}}{list of font attributes}
}}

\seealso{
\code{\link{add_col_title}}
}
\author{
Alicia Schep
}
\keyword{internal}
