% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ContinuousColorbar-class}
\alias{ContinuousColorbar-class}
\alias{ContinuousColorbar}
\title{ContinuousColorbar}
\description{
Class for storing continuous colorbar information
}
\section{Slots}{

\describe{
\item{\code{title}}{title for colorbar}

\item{\code{position}}{integer indicating relative position of colorbar}

\item{\code{colors}}{name of color palette or vector of colors}

\item{\code{zmid}}{midpoint of colorbar}

\item{\code{zmin}}{min of colorbar}

\item{\code{zmax}}{max of colorbar}
}}

\author{
Alicia Schep
}
\keyword{internal}
