% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Dendrogram-class}
\alias{Dendrogram-class}
\alias{Dendrogram}
\title{Dendrogram}
\description{
Class for storing dendrogram
}
\section{Slots}{

\describe{
\item{\code{xaxis}}{name of xaxis}

\item{\code{yaxis}}{name of yaxis}

\item{\code{data}}{hclust object}

\item{\code{side}}{side of plot on which dendrogram is positioned, controls 
orientation}
}}

\seealso{
\code{\link{add_row_dendro}}, \code{\link{add_col_dendro}},
\code{\link{add_row_clustering}}, \code{\link{add_col_clustering}}
}
\author{
Alicia Schep
}
\keyword{internal}
