% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DiscreteColorbar-class}
\alias{DiscreteColorbar-class}
\alias{DiscreteColorbar}
\title{DiscreteColorbar}
\description{
Class for storing discrete colorbar information
}
\section{Slots}{

\describe{
\item{\code{title}}{title for colorbar}

\item{\code{position}}{integer indicating relative position of colorbar}

\item{\code{colors}}{name of color palette or vector of colors}

\item{\code{ticktext}}{labels for categories}

\item{\code{tickvals}}{integer values for categories}
}}

\author{
Alicia Schep
}
\keyword{internal}
