% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GenericPlot-class}
\alias{GenericPlot-class}
\alias{GenericPlot}
\title{GenericPlot}
\description{
Class for storing an arbitrary subplot
}
\section{Slots}{

\describe{
\item{\code{xaxis}}{name of xaxis}

\item{\code{yaxis}}{name of yaxis}

\item{\code{data}}{list of plotly parameters}
}}

\seealso{
\code{\link{add_subplot}}
}
\author{
Alicia Schep
}
\keyword{internal}
