% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Iheatmap-class}
\alias{Iheatmap-class}
\alias{IheatmapHorizontal}
\alias{IheatmapVertical}
\alias{IheatmapHorizontal-class}
\alias{IheatmapVertical-class}
\title{Iheatmap-class}
\description{
Class to store complex interactive heatmap objects from iheatmapr package
}
\details{
This is a virtual class with two children classes, 
 IheatmapHorizontal and IheatmapVertical. For IheatmapHorizontal additional
 main heatmaps are added horizontally, and for IheatmapVertical additional
 main heatmaps are added vertically. For details on accessing certain slots 
 of this class, see \code{\link{access_component}} documentation.
}
\section{Slots}{

\describe{
\item{\code{plots}}{list of plot element in \code{\link{IheatmapPlots}} format}

\item{\code{shapes}}{list of shape element in \code{\link{IheatmapShapes}} format}

\item{\code{annotations}}{list of annotation elements in 
\code{\link{IheatmapAnnotations}} format}

\item{\code{xaxes}}{list of x axes in \code{\link{IheatmapAxes}} format}

\item{\code{yaxes}}{list of y axes in \code{\link{IheatmapAxes}} format}

\item{\code{colorbars}}{list of colorbars in \code{\link{IheatmapColorbars}} format}

\item{\code{colorbar_grid}}{colorbar grid parameters in
\code{\link{IheatmapColorbarGrid}} format}

\item{\code{current_xaxis}}{name of current x axis}

\item{\code{current_yaxis}}{name of current y axis}

\item{\code{layout}}{list of plotly layout parameters}
}}

\seealso{
\code{\link{iheatmap}}, \code{\link{main_heatmap}}, 
\code{\link{access_component}}
}
\author{
Alicia Schep
}
