% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iheatmap.R
\docType{methods}
\name{iheatmap}
\alias{iheatmap}
\alias{iheatmap,matrix-method}
\title{iheatmap}
\usage{
\S4method{iheatmap}{matrix}(data, x = default_x(data), y = default_y(data),
  cluster_rows = c("none", "hclust", "kmeans"), cluster_cols = c("none",
  "hclust", "kmeans"), row_clusters = NULL, col_clusters = NULL,
  row_k = NULL, col_k = NULL, row_clust_dist = stats::dist,
  col_clust_dist = stats::dist, name = "Signal", scale = c("none", "rows",
  "cols"), scale_method = c("standardize", "center", "normalize"),
  colors = NULL, col_clusters_colors = NULL,
  col_clusters_name = "Col<br>Clusters", row_clusters_colors = NULL,
  row_clusters_name = "Row<br>Clusters", show_row_clusters_colorbar = TRUE,
  show_col_clusters_colorbar = TRUE, row_annotation = NULL,
  col_annotation = NULL, row_annotation_colors = NULL,
  col_annotation_colors = NULL, row_labels = NULL, col_labels = NULL,
  row_title = NULL, col_title = NULL,
  colorbar_grid = setup_colorbar_grid(), font = list(), ...)
}
\arguments{
\item{data}{matrix of values to be plotted as heatmap}

\item{x}{x xaxis labels, by default colnames of data}

\item{y}{y axis labels, by default rownames of data}

\item{cluster_rows}{"none","hclust", or "k-means" for no clustering, 
hierarchical clustering, and k-means clustering of rows respectively}

\item{cluster_cols}{"none","hclust", or "k-means" for no clustering, 
hierarchical clustering, and k-means clustering of columnsrespectively}

\item{row_clusters}{vector of pre-determined row cluster assignment}

\item{col_clusters}{vector of pre-determined column cluster assignment}

\item{row_k}{number of clusters for rows, needed if cluster_rows is kmeans or 
optional if hclust}

\item{col_k}{number of clusters for columns, needed if cluster_rows is kmeans
or optional if hclust}

\item{row_clust_dist}{distance function to use for row clustering if 
hierarchical clustering}

\item{col_clust_dist}{distance function to use for column clustering 
if hierarchical clustering}

\item{name}{Name for colorbar}

\item{scale}{scale matrix by rows, cols or none}

\item{scale_method}{what method to use for scaling, either none, standardize, 
center, normalize}

\item{colors}{name of RColorBrewer palette or vector of colors for main 
heatmap}

\item{col_clusters_colors}{colors for col clusters annotation heatmap}

\item{col_clusters_name}{name for col clusters colorbar}

\item{row_clusters_colors}{colors for row clusters annotation heatmap}

\item{row_clusters_name}{name for row clusters colorbar}

\item{show_row_clusters_colorbar}{show the colorbar for row clusters?}

\item{show_col_clusters_colorbar}{show the colorbar for column clusters?}

\item{row_annotation}{row annotation data.frame}

\item{col_annotation}{column annotation data.frame}

\item{row_annotation_colors}{list of colors for row annotations heatmap}

\item{col_annotation_colors}{list of colors for col annotations heatmap}

\item{row_labels}{axis labels for y axis}

\item{col_labels}{axis labels for x axis}

\item{row_title}{x axis title}

\item{col_title}{y axis title}

\item{colorbar_grid}{colorbar grid parameters, should be result from 
\code{\link{setup_colorbar_grid}}}

\item{font}{list of font attributes}

\item{...}{additional argument to iheatmap}
}
\value{
\code{\link{Iheatmap-class}} object, which can be printed to generate 
an interactive graphic
}
\description{
iheatmap
}
\examples{
mat <- matrix(rnorm(24), nrow = 6)
annotation = data.frame(gender = c(rep("M", 3),rep("F",3)),
 age = c(20,34,27,19,23,30))
iheatmap(mat, 
 cluster_rows = "hclust",
 cluster_cols = "kmeans", 
 col_k = 3, 
 row_annotation = annotation)
}
\seealso{
\code{\link{iheatmap}}, \code{\link{add_iheatmap}}, 
\code{\link{as_plotly}}
}
\author{
Alicia Schep
}
