% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_utils.R
\docType{methods}
\name{iheatmap_list_utils}
\alias{iheatmap_list_utils}
\alias{length,IheatmapList-method}
\alias{as.list,IheatmapList-method}
\alias{names,IheatmapList-method}
\alias{`names<-`,IheatmapList-method}
\alias{`$`,IheatmapList-method}
\alias{`$<-`,IheatmapList-method}
\alias{`[`,IheatmapList-method}
\alias{`[<-`,IheatmapList-method}
\alias{`[[`,IheatmapList-method}
\alias{`[[<-`,IheatmapList-method}
\alias{lapply,IheatmapList-method}
\alias{vapply,IheatmapList-method}
\alias{[,IheatmapList-method}
\alias{[<-,IheatmapList-method}
\alias{[[,IheatmapList-method}
\alias{[[<-,IheatmapList-method}
\alias{$,IheatmapList-method}
\alias{$<-,IheatmapList-method}
\alias{names<-,IheatmapList-method}
\title{S4 List Utils for Iheatmap classes}
\usage{
\S4method{length}{IheatmapList}(x)

\S4method{as.list}{IheatmapList}(x)

\S4method{[}{IheatmapList}(x, i)

\S4method{[}{IheatmapList}(x, i) <- value

\S4method{[[}{IheatmapList}(x, i)

\S4method{[[}{IheatmapList}(x, i) <- value

\S4method{$}{IheatmapList}(x, name)

\S4method{$}{IheatmapList}(x, name) <- value

\S4method{names}{IheatmapList}(x)

\S4method{names}{IheatmapList}(x) <- value

\S4method{lapply}{IheatmapList}(X, FUN, ...)

\S4method{vapply}{IheatmapList}(X, FUN, FUN.VALUE, ..., USE.NAMES = TRUE)
}
\arguments{
\item{x}{input}

\item{FUN}{function to apply to each element of x}

\item{...}{additional arguments}

\item{FUN.VALUE}{template for return value from FUN}

\item{USE.NAMES}{logical, use names?}
}
\description{
These are utility methods for list-like classes in the package.
}
\keyword{internal}
