% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IheatmapAnnotation-class}
\alias{IheatmapAnnotation-class}
\alias{IheatmapAnnotation}
\title{IheatmapAnnotation}
\description{
Virtual class for storing annotation objects
}
\section{Slots}{

\describe{
\item{\code{xaxis}}{name of xaxis}

\item{\code{yaxis}}{name of yaxis}

\item{\code{data}}{main annotation data}
}}

\section{SubClasses}{

\itemize{
\item \code{\link{RowTitle-class}}
\item \code{\link{ColumnTitle-class}}
\item \code{\link{RowLabels-class}}
\item \code{\link{ColumnLabels-class}}
}
}

\author{
Alicia Schep
}
\keyword{internal}
