% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IheatmapAnnotations-class}
\alias{IheatmapAnnotations-class}
\alias{IheatmapAnnotations}
\title{IheatmapAnnotations}
\description{
Class for storing \code{\link{IheatmapAnnotation}} objects
}
\seealso{
\code{\link{IheatmapAnnotation-class}}, \code{\link{Iheatmap-class}}
}
\keyword{internal}
