% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IheatmapAxes-class}
\alias{IheatmapAxes-class}
\alias{IheatmapAxes}
\title{IheatmapAxes}
\description{
Class for storing \code{\link{IheatmapAxis}} objects
}
\section{Slots}{

\describe{
\item{\code{axis}}{x or y?}
}}

\seealso{
\code{\link{IheatmapAxis-class}}, \code{\link{Iheatmap-class}}
}
\author{
Alicia Schep
}
\keyword{internal}
