% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IheatmapAxis-class}
\alias{IheatmapAxis-class}
\alias{IheatmapAxis}
\title{IheatmapAxis}
\description{
Class for storing axis information
}
\section{Slots}{

\describe{
\item{\code{id}}{plotly id for axis}

\item{\code{domain_start}}{start of domain (0 to 1)}

\item{\code{domain_end}}{end of domain (0 to 1)}

\item{\code{anchor}}{anchor for axis}

\item{\code{layout}}{plotly layout parameters}
}}

\section{SubClasses}{

\itemize{
\item \code{\link{IheatmapX-class}}
\item \code{\link{IheatmapY-class}}
\item \code{\link{IheatmapMainX-class}}
\item \code{\link{IheatmapMainY-class}} 
}
}

\seealso{
\code{\link{Iheatmap-class}}
}
\author{
Alicia Schep
}
\keyword{internal}
