% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IheatmapColorbar-class}
\alias{IheatmapColorbar-class}
\alias{IheatmapColorbar}
\title{IheatmapColorbar}
\description{
Virtual class for storing colorbar objects
}
\section{Slots}{

\describe{
\item{\code{title}}{title for colorbar}

\item{\code{position}}{integer indicating relative position of colorbar}

\item{\code{colors}}{name of color palette or vector of colors}
}}

\section{SubClasses}{

\itemize{
\item \code{\link{ContinuousColorbar-class}}
\item \code{\link{DiscreteColorbar-class}}
}
}

\author{
Alicia Schep
}
\keyword{internal}
