% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IheatmapColorbarGrid-class}
\alias{IheatmapColorbarGrid-class}
\alias{IheatmapColorbarGrid}
\title{IheatmapColorbarGrid}
\description{
Parameters for setting up colorbars
}
\section{Slots}{

\describe{
\item{\code{nrows}}{number of rows}

\item{\code{x_spacing}}{spacing between colorbars horizontally}

\item{\code{y_spacing}}{spacing between colorbars vertically}

\item{\code{y_length}}{length of colorbars vertically}

\item{\code{x_start}}{start position horizontally}

\item{\code{y_start}}{start position vertically}
}}

\seealso{
\code{\link{setup_colorbar_grid}}, \code{\link{Iheatmap-class}}
}
\author{
Alicia Schep
}
\keyword{internal}
