% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IheatmapColorbars-class}
\alias{IheatmapColorbars-class}
\alias{IheatmapColorbars}
\title{IheatmapColorbars}
\description{
Class for storing \code{\link{IheatmapColorbar}} objects
}
\seealso{
\code{\link{IheatmapColorbar-class}}, \code{\link{Iheatmap-class}}
}
\author{
Alicia Schep
}
\keyword{internal}
