% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IheatmapMainX-class}
\alias{IheatmapMainX-class}
\alias{IheatmapMainX}
\title{IheatmapMainX}
\description{
Class for storing X axis information for "main" x axis-- x axis for 
\code{\link{MainHeatmap-class}}
}
\section{Slots}{

\describe{
\item{\code{id}}{plotly id for axis}

\item{\code{domain_start}}{start of domain (0 to 1)}

\item{\code{domain_end}}{end of domain (0 to 1)}

\item{\code{anchor}}{anchor for axis}

\item{\code{layout}}{plotly layout parameters}

\item{\code{categorical}}{is axis categorical?}

\item{\code{order}}{ordering of columns}

\item{\code{text}}{text labels for columns}
}}

\seealso{
\code{\link{Iheatmap-class}}, \code{\link{IheatmapAxis}}
}
\author{
Alicia Schep
}
\keyword{internal}
