% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IheatmapPlot-class}
\alias{IheatmapPlot-class}
\alias{IheatmapPlot}
\title{IheatmapPlot}
\description{
Virtual class for storing plot objects
}
\section{Slots}{

\describe{
\item{\code{xaxis}}{name of xaxis}

\item{\code{yaxis}}{name of yaxis}

\item{\code{data}}{main plot data}
}}

\section{SubClasses}{

\itemize{
\item \code{\link{MainHeatmap-class}}
\item \code{\link{RowAnnotation-class}}
\item \code{\link{ColumnAnnotation-class}}
\item \code{\link{RowPlot-class}}
\item \code{\link{ColumnPlot-class}}
\item \code{\link{GenericPlot-class}}
}
}

\author{
Alicia Schep
}
\keyword{internal}
