% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IheatmapPlots-class}
\alias{IheatmapPlots-class}
\alias{IheatmapPlots}
\title{IheatmapPlots}
\description{
Class for storing \code{\link{IheatmapPlot}} objects
}
\seealso{
\code{\link{IheatmapPlot-class}}, \code{\link{Iheatmap-class}}
}
\author{
Alicia Schep
}
\keyword{internal}
