% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{iheatmapr_event}
\alias{iheatmapr_event}
\title{Access iheatmapr user input event data in shiny}
\usage{
iheatmapr_event(
  object,
  event = c("hover", "click", "relayout"),
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{object}{\code{\link{Iheatmap-class}} object}

\item{event}{The type of plotly event. Currently 'plotly_hover',
'plotly_click', 'plotly_selected', and 'plotly_relayout' are supported.}

\item{session}{a shiny session object (the default should almost always be used).}
}
\description{
This function must be called within a reactive shiny context.
}
\examples{
\dontrun{
shiny::runApp(system.file("examples", "shiny_example", package = "iheatmapr"))
}
}
