% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IheatmapShape-class}
\alias{IheatmapShape-class}
\alias{IheatmapShape}
\title{IheatmapShape}
\description{
Virtual class for storing shape objects
}
\section{Slots}{

\describe{
\item{\code{xaxis}}{name of xaxis}

\item{\code{yaxis}}{name of yaxis}

\item{\code{data}}{main shape data}
}}

\section{SubClasses}{

\itemize{
\item \code{\link{Dendrogram-class}}
}
}

\author{
Alicia Schep
}
\keyword{internal}
