% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IheatmapShapes-class}
\alias{IheatmapShapes-class}
\alias{IheatmapShapes}
\title{IheatmapShapes}
\description{
Class for storing \code{\link{IheatmapShapes}} objects
}
\seealso{
\code{\link{IheatmapShape-class}}, \code{\link{Iheatmap-class}}
}
\author{
Alicia Schep
}
\keyword{internal}
