% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IheatmapX-class}
\alias{IheatmapX-class}
\alias{IheatmapX}
\title{IheatmapX}
\description{
Class for storing X axis information
}
\section{Slots}{

\describe{
\item{\code{id}}{plotly id for axis}

\item{\code{domain_start}}{start of domain (0 to 1)}

\item{\code{domain_end}}{end of domain (0 to 1)}

\item{\code{anchor}}{anchor for axis}

\item{\code{layout}}{plotly layout parameters}
}}

\seealso{
\code{\link{Iheatmap-class}}, \code{\link{IheatmapAxis}}
}
\author{
Alicia Schep
}
\keyword{internal}
