% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_widget.R
\name{knit_print.Iheatmap}
\alias{knit_print.Iheatmap}
\title{knit_print.Iheatmap}
\usage{
\method{knit_print}{Iheatmap}(x, ...)
}
\arguments{
\item{x}{Iheatmap object}

\item{...}{Passed to \code{\link[knitr]{knit_print}}.}
}
\description{
knit_print.Iheatmap
}
\keyword{internal}
