% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MainHeatmap-class}
\alias{MainHeatmap-class}
\alias{MainHeatmap}
\title{MainHeatmap}
\description{
Class for storing main heatmap
}
\section{Slots}{

\describe{
\item{\code{xaxis}}{name of xaxis}

\item{\code{yaxis}}{name of yaxis}

\item{\code{data}}{matrix of heatmap values}

\item{\code{colorbar}}{name of colorbar}

\item{\code{show_colorbar}}{show the colorbar?}
}}

\seealso{
\code{\link{main_heatmap}}
}
\author{
Alicia Schep
}
\keyword{internal}
