% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{make_component}
\alias{make_component}
\alias{make_trace,MainHeatmap-method}
\alias{make_trace,RowAnnotation-method}
\alias{make_trace,ColumnAnnotation-method}
\alias{make_trace,RowPlot-method}
\alias{make_trace,ColumnPlot-method}
\alias{make_trace,GenericPlot-method}
\alias{make_shapes,Dendrogram-method}
\alias{make_annotations,RowTitle-method}
\alias{make_annotations,ColumnTitle-method}
\alias{make_annotations,RowLabels-method}
\alias{make_annotations,ColumnLabels-method}
\alias{make_colorbar,ContinuousColorbar,IheatmapColorbarGrid-method}
\alias{make_colorbar,DiscreteColorbar,IheatmapColorbarGrid-method}
\alias{make_trace}
\alias{make_shapes}
\alias{make_annotations}
\alias{make_colorbar}
\title{Convert iheatmapr subcomponents to plotly format}
\usage{
make_trace(x, ...)

make_shapes(x, ...)

make_annotations(x, ...)

make_colorbar(cb, grid)
}
\arguments{
\item{x}{\code{\link{IheatmapPlot-class}}, \code{\link{IheatmapShape-class}},
or \code{\link{IheatmapAnnotation-class}} object}

\item{...}{additional arguments specific to component}
}
\description{
These are generic methods for converting \code{link{Iheatmap-class}}  plot 
components to plotly lists. Not intended for end users; exported for
developers seeking to create new Iheatmap subplots. Any new 
\code{link{IheatmapPlot}}, \code{link{IheatmapShape}},
 \code{link{IheatmapAnnotation}}, or \code{link{IheatmapColorbar}} child class
 should have one of these methods.
}
\keyword{internal}
