% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RowPlot-class}
\alias{RowPlot-class}
\alias{RowPlot}
\title{RowPlot}
\description{
Class for storing generic row plot
}
\section{Slots}{

\describe{
\item{\code{xaxis}}{name of xaxis}

\item{\code{yaxis}}{name of yaxis}

\item{\code{data}}{vector of values}

\item{\code{name}}{tracename}

\item{\code{type}}{trace type}

\item{\code{showlegend}}{show the legend?}

\item{\code{additional}}{additional plotly parameters}
}}

\seealso{
\code{\link{add_row_plot}}, \code{\link{add_row_barplot}},
\code{\link{add_row_summary}}
}
\author{
Alicia Schep
}
\keyword{internal}
