% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_plotly.R
\docType{methods}
\name{save_iheatmap}
\alias{save_iheatmap}
\alias{save_iheatmap,Iheatmap,character-method}
\title{save_iheatmap}
\usage{
\S4method{save_iheatmap}{Iheatmap,character}(p, filename, ...)
}
\arguments{
\item{p}{\code{link{Iheatmap-class}} object}

\item{filename}{name of file}

\item{...}{additional arguments to \code{\link[htmlwidgets]{saveWidget}} for 
saving as html or \code{\link[plotly]{export}} for saving as pdf/png/jpeg}
}
\description{
save an \code{link{Iheatmap-class}} object, either as standalone HTML or as static 
pdf/png/jpeg
}
\examples{
mat <- matrix(rnorm(24), nrow = 6)
hm <- iheatmap(mat)
\dontrun{
save_iheatmap(hm)
}
}
\author{
Alicia Schep
}
