% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_test.R
\name{test_iheatmapr_event}
\alias{test_iheatmapr_event}
\title{test_iheatmapr_event}
\usage{
test_iheatmapr_event(ihm, event = c("click", "hover", "relayout"))
}
\arguments{
\item{ihm}{Iheatmap object}

\item{event}{name of event, either "click","hover", or "relayout"}
}
\value{
shiny app
}
\description{
test_iheatmapr_event
}
\examples{

\dontrun{
  mat <- matrix(rnorm(20), ncol = 5, nrow = 4)  
  hm <- main_heatmap(mat) 
  test_iheatmapr_event(hm, "click")
}
}
