% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_widget.R
\name{to_plotly}
\alias{to_plotly}
\alias{to_plotly_list}
\alias{to_plotly_json}
\title{Convert Iheatmap to plotly spec}
\usage{
to_plotly_list(p)

to_plotly_json(p)
}
\arguments{
\item{p}{\code{\link{Iheatmap-class}} object to convert}
}
\value{
Returns a JSON for a plotly spec for to_plotly_spec and
as a list of same plotly object for to_plotly_list.
}
\description{
Function  to convert \code{link{Iheatmap-class}} object to a plotly spec 
either as a list or json
}
\examples{

mat <- matrix(rnorm(24), nrow = 6)
hm_json <- iheatmap(mat) \%>\% to_plotly_json()
hm_list <- iheatmap(mat) \%>\% to_plotly_list()
}
