% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_widget.R
\name{to_widget}
\alias{to_widget}
\alias{to_widget,Iheatmap-method}
\title{to_widget}
\usage{
\S4method{to_widget}{Iheatmap}(p)
}
\arguments{
\item{p}{\code{\link{Iheatmap-class}} object to convert}
}
\value{
htmlwidgets object
}
\description{
Function to convert \code{link{Iheatmap-class}} object to widget object
}
\examples{

mat <- matrix(rnorm(24), nrow = 6)
hm <- iheatmap(mat) \%>\% to_widget()
class(hm)

# Print heatmap if interactive session 
if (interactive()) hm 
}
\seealso{
\code{\link{iheatmap}}, \code{\link{main_heatmap}}
}
\author{
Alicia Schep
}
