% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_plotly.R
\name{knit_print.Iheatmap}
\alias{knit_print.Iheatmap}
\title{knit_print.Iheatmap}
\usage{
\method{knit_print}{Iheatmap}(x, options)
}
\arguments{
\item{x}{Iheatmap object}

\item{options}{knitr options}
}
\description{
knit_print.Iheatmap
}
\keyword{internal}
